Refactor: JWT refresh, logout 함수 유틸로 분리
