+++ src/app/main/_header.tsx
@@ -6,13 +6,8 @@ import { FaUser } from 'react-icons/fa';
 import { userProfileMeDto } from '../_dto/fetch-profile/Profile.dto';
 import DialogModalTwoButton from '../_components/modalTwoButton';
 import DialogModalOneButton from '../_components/modalOneButton';
-import { RefreshTokenReqDto } from '../_dto/refresh-token/refresh-token.dto';
-
-const logout = async () => {
-  await fetch('/api/web/logout');
-  localStorage.removeItem('user_handle');
-  window.location.reload();
-};
+import { refreshJwt } from '@/utils/refreshJwt/refresh-jwt-token';
+import { logout } from '@/utils/logout/logout';
 
 export default function MainHeader() {
   const [user, setUser] = useState<userProfileMeDto>();
@@ -21,7 +16,6 @@ export default function MainHeader() {
 
   const fetchMyProfile = async () => {
     const user_handle = localStorage.getItem('user_handle');
-    const last_token_refresh = Number.parseInt(localStorage.getItem('last_token_refresh') ?? '0');
     const now = Math.ceil(Date.now() / 1000);
 
     if (user_handle) {
@@ -34,28 +28,10 @@ export default function MainHeader() {
         }
         return;
       }
-      // FIXME: 3600초 대신 적당한 시간으로 고치기
+      // JWT 리프레시로부터 1시간이 지난 경우 refresh 시도
+      const last_token_refresh = Number.parseInt(localStorage.getItem('last_token_refresh') ?? '0');
       if (now - last_token_refresh > 3600) {
-        localStorage.setItem('last_token_refresh', `${now}`);
-        try {
-          const req: RefreshTokenReqDto = {
-            handle: user_handle,
-            last_refreshed_time: last_token_refresh,
-          }
-          const res = await fetch('/api/web/refresh-token', {
-            method: 'POST',
-            headers: {
-              'content-type': 'application/json'
-            },
-            body: JSON.stringify(req),
-          });
-          if (res.status === 401 || res.status === 403) {
-            alert(`마스토돈/미스키 에서 앱 인증이 해제되었어요! ${await res.text()}`);
-            await logout();
-          }
-        } catch (err) {
-          console.error(err);
-        }
+        await refreshJwt();
       }
       const data = await res.json();
       return data;
